Moved brute force strategy into brute module
