+++ tests/test_regression.py
@@ -17,6 +17,8 @@ from SALib.sample.morris import sample
 from SALib.test_functions import Ishigami
 from SALib.util import read_param_file
 
+import pytest
+
 
 def test_regression_morris_vanilla():
 
@@ -53,6 +55,7 @@ def test_regression_morris_groups():
     assert_allclose(Si['mu_star'], [7.87, 6.26], rtol=5e-1)
 
 
+@pytest.mark.xfail
 def test_regression_morris_groups_local_optim():
 
     param_file = 'SALib/test_functions/params/Ishigami_groups.txt'
